services/auth/pyproject.toml
[tool.poetry]
name = "fcc-microservices-auth"
version = "0.1.0"
description = ""
authors = ["DimaK <korbanitskiy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
email-validator = "^2.1.1"
httpx = "^0.27.0"
pyyaml = "^6.0.1"
uvicorn = "^0.27.1"
orjson = "^3.9.15"
ormar = "^0.12.2"
asyncpg = "^0.29.0"
tenacity = "^8.2.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
pika = "^1.3.2"
pymongo = "^4.6.2"
pytest = "^8.0.2"
pytest-env = "^1.1.3"
pytest-asyncio = "0.21.0"
factory-boy = "^3.3.0"
pytest-mock = "^3.12.0"
pytest-xdist = "^3.5.0"
freezegun = "^1.4.0"
databases = {version = "0.6.2", extras = ["postgresql"]}
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
